+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # VocalFusion Raspberry Pi Setup Change Log
 
+## 2.0.0
+
+  * Use default I2C hardware
+  * Add installation of Py3 packages
+
 ## 1.3.2
 
   * Added audacity installation and setup for XVF3510